+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Todo Tree Change Log
 
+- Fix compact folders when grouping by tag
+
 ## v0.0.195 - 2021-01-20
 
 - Fix icon colour check